+++ src/conline.cpp
@@ -49,7 +49,7 @@ void conline::set_defaults()
 
     R0.zeros(D, P, K);
 
-    predictions_got_sorted.zeros(T + T_E_Y);
+    predictions_got_sorted.zeros(T + T_E_Y, D);
 }
 
 void conline::set_grid_objects()
@@ -218,7 +218,7 @@ void conline::learn()
                 {
                     if ((x == opt_index(t)) && (!tmp_preds_vec.is_sorted()))
                     {
-                        predictions_got_sorted(t) = 1;
+                        predictions_got_sorted(t, d) = 1;
                     }
                     tmp_preds_vec = arma::sort(tmp_preds_vec, "ascend", 0);
                 }
@@ -505,6 +505,10 @@ void conline::learn()
             // Sort predictions if quantile_crossing is prohibited
             if (!allow_quantile_crossing)
             {
+                if (!tmp_preds_vec.is_sorted())
+                {
+                    predictions_got_sorted(t, d) = 1;
+                }
                 tmp_preds_vec = arma::sort(tmp_preds_vec, "ascend", 0);
             }
             predictions.tube(t, d) = tmp_preds_vec;
@@ -665,8 +669,8 @@ void conline::init_update(
     // // Output Objects
     predictions = Rcpp::as<arma::cube>(object["predictions"]);
     predictions.resize(T + T_E_Y, D, P);
-    predictions_got_sorted = Rcpp::as<arma::vec>(object["predictions_got_sorted"]);
-    predictions_got_sorted.resize(T + T_E_Y);
+    predictions_got_sorted = Rcpp::as<arma::mat>(object["predictions_got_sorted"]);
+    predictions_got_sorted.resize(T + T_E_Y, D);
     weights.set_size(T + 1);
     weights.rows(0, start) = Rcpp::as<arma::field<cube>>(object["weights"]);